Refactored out the withSearchBar - stage 2.
